+++ api/changelog.md
@@ -3,6 +3,8 @@
 ## v0.4.1
 * rewrite undici fetch integration.
 * add debug parameter.
+* add cause in Exception
+* remove console.log
 
 ## v0.4.0
 * use undici fetch (will fix issue 14 in the next release)